+++ index.php
@@ -8,7 +8,7 @@
  * @license GPL-3, http://www.gnu.org/licenses/
  */
 
-include_once 'php/backend.php';
+include_once 'assets/php/backend.php';
 
 ?><!DOCTYPE html>
 <html>
@@ -18,13 +18,13 @@ include_once 'php/backend.php';
 		<meta name="apple-mobile-web-app-capable" content="yes"/>
 		<meta name="viewport" content="user-scalable=no, width=device-width, initial-scale=1.0, maximum-scale=1.0"/>
 		<meta name="csrf-token" content="<?php echo $_SESSION['csrf_token'];?>">
-		<link rel="stylesheet" type="text/css" href="css/style.min.css"/>
-		<link rel="icon" type="image/png" href="img/ship.png"/>
-		<script src="js/vendor/jquery.min.js"></script>
-		<script src="js/vendor/starscroll.min.js"></script>
-		<script src="js/vendor/base64.min.js"></script>
-		<script src="js/vendor/wade.min.js"></script>
-		<script src="js/init.min.js"></script>
+		<link rel="stylesheet" type="text/css" href="assets/css/style.min.css"/>
+		<link rel="icon" type="image/png" href="assets/img/ship.png"/>
+		<script src="assets/js/vendor/jquery.min.js"></script>
+		<script src="assets/js/vendor/starscroll.min.js"></script>
+		<script src="assets/js/vendor/base64.min.js"></script>
+		<script src="assets/js/vendor/wade.min.js"></script>
+		<script src="assets/js/init.min.js"></script>
 		<title>Overkill - In space no one hears you cry</title>
 	</head>
 	<body>
@@ -44,7 +44,7 @@ include_once 'php/backend.php';
 					</thead>
 					<tbody id="highscoreTable">
 						<tr>
-							<td colspan="2" class="img"><img src="img/loading.svg" alt=""/></td>
+							<td colspan="2" class="img"><img src="assets/img/loading.svg" alt=""/></td>
 						</tr>
 					</tbody>
 				</table>